# feat: establish performance baseline measurements for global state elimination